autopilot: add stop ctx
